refactored after POJO learning
